Debug: Add detailed console logging for victory event flow
- Added logging to OnVictoryAchieved in GameCanvas
- Added logging to GoalReached event handler in Home page
- Added logging to EndGameWithVictory method
- Wrapped OnVictoryAchieved in InvokeAsync for thread safety
- This will help diagnose why victory screen doesn't show despite console logging victory

+++ backend/tests/Po.PoDropSquare.E2E.Tests/VictoryCountdownE2ETests.cs
@@ -167,10 +167,10 @@ public class VictoryCountdownE2ETests : IAsyncLifetime
 
         // Now drop more blocks to potentially make tower unstable
         Console.WriteLine("🎮 Dropping additional blocks to destabilize tower...");
-        var sideClickX = boundingBox.X + boundingBox.Width * 0.3; // Click to the side
+        var sideClickX = (float)(boundingBox.X + boundingBox.Width * 0.3); // Click to the side
         for (int i = 0; i < 5; i++)
         {
-            await _page.Mouse.ClickAsync(sideClickX, clickY);
+            await _page.Mouse.ClickAsync(sideClickX, (float)clickY);
             await Task.Delay(200);
         }
